feat: 分离下载功能函数
